requirements.txt
python-dotenv
pandas
google-generativeai
python-docx
schedule
requests 